More tweaks from comments
