+++ src/Libraries/RevitNodes/Properties/Resources.en-US.resx
@@ -312,4 +312,10 @@
   <data name="MultipleSurfacesIntroducedAfterConversion" xml:space="preserve">
     <value>There are multiple surfaces converted from the input face.</value>
   </data>
+  <data name="DynamoSurfaceToRevitBRepFailure" xml:space="preserve">
+    <value>An unexpected failure occurred when attempting to convert the surface into a Revit BRep</value>
+  </data>
+  <data name="DynamoSolidToRevitBRepFailure" xml:space="preserve">
+    <value>An unexpected failure occurred when attempting to convert the solid into a Revit BRep</value>
+  </data>
 </root>
